Blessed Scan
Blessed Scan is an advanced network management and analysis tool designed for IT professionals and cybersecurity experts. It offers a range of powerful features to enhance network security and analysis capabilities:

URL to IP: This feature allows you to convert any URL into its corresponding IP address, aiding in network analysis and security assessments.

Fast Scan: Blessed San provides a fast scanning function to quickly assess network systems and services. This feature is useful for a quick overview of open ports and active services on the target system.

Deep Scan: The deep scan functionality offers a comprehensive and detailed examination of networks and systems. It helps in in-depth port scanning, service enumeration, and vulnerability detection.

WHOIS Lookup: This feature enables you to retrieve detailed information about domain owners, including contact details and registration data, which is valuable for security investigations and network analysis.

Overall, Blessed San is a versatile and effective tool for those seeking advanced solutions for network analysis and security enhancement.